HV: separate const dmar table definition from sbl
const dmar table is used for sbl board, and it can be
a single/independable file.

Tracked-On: #1842
Acked-by: Eddie Dong <eddie.dong@intel.com>
